+++ src/mo_frontend/source_token.ml
@@ -62,8 +62,6 @@ type token =
   | BANG
   | AND
   | OR
-  | IMPLIES
-  | OLD
   | NOT
   | IMPORT
   | INCLUDE
@@ -129,7 +127,6 @@ type token =
   | UNDERSCORE
   | WEAK
   | COMPOSITE
-  | INVARIANT
   (* Trivia *)
   | LINEFEED of line_feed
   | SINGLESPACE
@@ -200,8 +197,6 @@ let to_parser_token :
   | BANG -> Ok Parser.BANG
   | AND -> Ok Parser.AND
   | OR -> Ok Parser.OR
-  | IMPLIES -> Ok Parser.IMPLIES
-  | OLD -> Ok Parser.OLD
   | NOT -> Ok Parser.NOT
   | IMPORT -> Ok Parser.IMPORT
   | INCLUDE -> Ok Parser.INCLUDE
@@ -265,7 +260,6 @@ let to_parser_token :
   | PRIM -> Ok Parser.PRIM
   | UNDERSCORE -> Ok Parser.UNDERSCORE
   | COMPOSITE -> Ok Parser.COMPOSITE
-  | INVARIANT -> Ok Parser.INVARIANT
   | PIPE -> Ok Parser.PIPE
   | WEAK -> Ok Parser.WEAK
   (*Trivia *)
@@ -404,9 +398,6 @@ let string_of_parser_token = function
   | Parser.PRIM -> "PRIM"
   | Parser.UNDERSCORE -> "UNDERSCORE"
   | Parser.COMPOSITE -> "COMPOSITE"
-  | Parser.INVARIANT -> "INVARIANT"
-  | Parser.IMPLIES -> "IMPLIES"
-  | Parser.OLD -> "OLD"
   | Parser.PIPE -> "PIPE"
   | Parser.WEAK -> "WEAK"